Fix compiler warning: comparison of integer expressions of different signedness

+++ src/memory/AmdFlash.cc
@@ -663,7 +663,7 @@ bool AmdFlash::checkCommandBufferProgram()
 			if (cmd[3].value < chip.program.pageSize && getSector(cmd[3].addr) == getSector(cmd[2].addr)) {
 				return true;
 			}
-		} else if (cmd.size() <= 5 + cmd[3].value) {
+		} else if (cmd.size() <= size_t(5 + cmd[3].value)) {
 			const size_t pageMask = ~(chip.program.pageSize - 1);
 			if ((cmd.back().addr & pageMask) == (cmd[4].addr & pageMask)) {
 				status = (status & 0x7F) | (~cmd.back().value & 0x80);
